Fix invalid linking against python library
Especially on macOS we should not link the coda python extension explicitly against the python shared library.
This conflicts with imports of coda that are done by a python executable that embeds a station version of the python library.
On Python 3.7 this produces segfaults due to thread states not being initialized (since it is picked up from the wrong library)