Have IDL/MATLAB interfaces return libcoda version as CODA version
Now that the IDL and MATLAB extensions link dynamically to the libcoda library an inconsistency can happen between the version of the library and the interface.
In that case it is currently more important to provide the version of the library. We may later look at a way to provide the CODA version of both components.