add 'type path', to pre-determine/cache lookups
similar to type-trees, we pre-determine information
based on the cursor nodetype, but now along the path to be
traversed.

for now we only pre-determine field lookups, as those are
the most expensive when reading scalars.

in the future, we could add more information to speed up
various other use cases (small arrays, non-scalar leafs..).